uudoc,uucore_procs: move md parsing to help_parser
