+++ src/help_parser/Cargo.toml
@@ -1,0 +1,5 @@
+[package]
+name = "help_parser"
+version = "0.0.18"
+edition = "2021"
+license = "MIT"
